+++ src/test/java/gov/usgs/wma/waterdata/collections/FeaturesParamsIT.java
@@ -49,6 +49,9 @@ public class FeaturesParamsIT extends BaseCollectionsIT  {
 		}
 	}
 
+
+
+
 	@Test
 	public void noPrevLinkTest() {
 		for (String collectionId : COLLECTION_IDS) {
@@ -144,6 +147,174 @@ public class FeaturesParamsIT extends BaseCollectionsIT  {
 			doJsonCompare(rtn, compareFile);
 		}
 	}
+
+
+
+	@Test
+	public void countryTest() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?country=MX&country=US",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		String compareFile = "featureCollection/monitoring-locations/countries_mx_us.json";
+		doJsonCompare(rtn, compareFile);
+	}
+
+	@Test
+	public void countryTestNotFound() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?country=MX",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
+	}
+
+	@Test
+	public void countryTestInvalid() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?country=MXUS",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.BAD_REQUEST));
+	}
+
+
+
+	@Test
+	public void stateTest() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?state=26",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		String compareFile = "featureCollection/monitoring-locations/states_mi.json";
+		doJsonCompare(rtn, compareFile);
+	}
+
+
+	@Test
+	public void stateTestNotFound() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?state=57",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
+	}
+
+
+	@Test
+	public void stateTestInvalid() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?state=57000",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.BAD_REQUEST));
+	}
+
+	@Test
+	public void countyTest() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?county=48205",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		String compareFile = "featureCollection/monitoring-locations/counties_tx_48205.json";
+		doJsonCompare(rtn, compareFile);
+	}
+
+	@Test
+	public void countyTestNotFound() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?county=06071",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
+	}
+
+	@Test
+	public void countyTestInvalid() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?county=Hartley",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.BAD_REQUEST));
+	}
+
+
+	@Test
+	public void hydrologicalUnitsTest() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?hydrologicalUnit=040103020107",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		String compareFile = "featureCollection/monitoring-locations/hydrological_unit_040103020107.json";
+		doJsonCompare(rtn, compareFile);
+	}
+
+	@Test
+	public void hydrologicalUnitsTestNotFound() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?hydrologicalUnit=000000000000",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
+	}
+
+
+	@Test
+	public void hydrologicalUnitsTestInvalid() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?hydrologicalUnit=INVALID",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.BAD_REQUEST));
+	}
+
+
+	@Test
+	public void monitoringLocationTypeTest() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?monitoringLocationType=Well",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		String compareFile = "featureCollection/monitoring-locations/monitoring_location_type_well.json";
+		doJsonCompare(rtn, compareFile);
+	}
+
+	@Test
+	public void monitoringLocationTypeTestNotFound() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?monitoringLocationType=Unknown",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
+	}
+
+	@Test
+	public void agencyCodeTest() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?agencyCode=USGS",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+	}
+
+	@Test
+	public void agencyCodeTestNotFound() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?agencyCode=Unknown",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
+	}
+
+	@Test
+	public void nationalAquiferCodeTest() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?nationalAquiferCode=N9999OTHER",
+			String.class);
+
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		String compareFile = "featureCollection/monitoring-locations/nat_aqfr_cd_N9999OTHER.json";
+		doJsonCompare(rtn, compareFile);
+
+	}
+
+	@Test
+	public void nationalAquiferCodeTestNotFound() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity("/collections/monitoring-locations/items?nationalAquiferCode=DOESNTEXIST",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
+	}
+
+	@Test
+	public void monitoringLocationNumberTest() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity(
+			"/collections/monitoring-locations/items?monitoringLocationNumber=USGS-343204093005501",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		String compareFile = "featureCollection/monitoring-locations/monitoring_location_number_USGS-343204093005501.json";
+		doJsonCompare(rtn, compareFile);
+	}
+
+	@Test
+	public void monitoringLocationNumberTestNotFound() throws IOException {
+		ResponseEntity<String> rtn = restTemplate.getForEntity(
+			"/collections/monitoring-locations/items?monitoringLocationNumber=DOESNTEXIST",
+			String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
+	}
+
 
 	private void doJsonCompare(ResponseEntity<String> rtn, String compareFile) {
 		try {