+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   Removed deprecated end point: /monitoring-location/{monitoringLocationId}/time-series/{timeSeriesId}
 -   Performance improvement on end point: /collections/{collectionId}/items/{featureId}/observations/statistical-time-series
 -   Unapproved time series data older than 1095 days is not returned from the services (USGS rule)
+-   Added filtering to /collections/{collectionId}/items
 
 ## [0.6.0](https://github.com/USGS/time-series-services/compare/time-series-services-0.5.0...time-series-services-0.6.0) - 2020-04-29
 ### Changed